Art implementation2

+++ Assets/Scripts/Combat_Refactored/Interfaces/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Hit(int DamageAmount);
+    void Hit(float DamageAmount);
 }
